correção de erros

+++ Grupo2A/Model/Refeicao.cs
@@ -11,7 +11,7 @@ namespace Cozinha_BE.Model
         public required long PratoId { get; set; } // Referência ao prato
         public required DateTime Data { get; set; } // Data da refeição
         [ForeignKey("TipoRefeicao")]
-        public required int TipoRefeicaoId { get; set; } // Referência ao tipo de refeição
+        public required long TipoRefeicaoId { get; set; } // Referência ao tipo de refeição
         public required int QuantidadeProduzida { get; set; } // Quantidade produzida
 
         public virtual required Prato Prato { get; set; } // Navegação para Prato
